Set `Depth` limit to `Int`
If `Depth` is larger than `maxBound` it will cycle back to the beginning
of the `Series`. In practice `Depth`s so huge are never used, but since
I'm going to introduce `Int8`, which will be based on the `Serial`
instance of `Int`, it makes sense to already handle a limit.

This also indirectly sets a limit to `Integer` where there is no integer
overflow, but the `maxBound :: Int` depth limit should be enough for
any practical situation.

See #30.